Update to Correlation and McClellan Osc
This is an update to determine the correlation of the stock to the rest
of the market. The idea behind this is to find stocks that perform well
when the rest of the market is performing well, or if the market is
performing poorly, buy stocks that are performing well against the
market.