Adding input for price to exit alerts
This is an alert that stocks is moving outside of normal trading range.
Does not necessarily mean that it is a hard exit on either side but it
is a signal to monitor much more closely.
Defaults are a 21 day RSI SR for a target, and a three day low/high for
a stoploss. If the target is reached, then switch to a 180 day RSI SR,
and a stoploss of a close below/above the previous day. Let profitable
breakouts ride until there's a hint of it stalling out.