MAJOR UPDATE: Swing strategy
The swing strategy is based on stochastic slow crossover ONLY. The only
confirmation for the strategy is a positive day (close > open). The goal
is to capture smaller price movements in order to increase win
percentage; while sacrificing bigger gains. Once a purchase is made, the
target a bit arbitrary and based upon the discretion of the trader. The
target will likely be the first point of resistance over the last 10
days, which will typically be about 2%-3% profits from purchase to
target.