Major: Combine GMMA/Stochastic/MACD
For a daily chart, looking for a price crossover on the entire set of
short range MAs. Selling on the stochastic 14,3 crossover. For a hourly
chart, looking for the short range MAs to crossover on the long range
MAs. Selling when the short range MAs short to converge and move
sideways.