Major: Going back to what worked before...
I went back to determine what I was doing about a year ago to what
worked really well. I lost my way for a while, but now I am using all
indicators in conjunction to improve probability of a good entry. I am
looking for a new high compared to yesterday, crossovers on RSI/RSI SMA,
Stochastic, and MACD, as well as heavy volume, and candlesticks. The
volume and candlesticks are not requirements, but they are def
recommended.